# openapi.json
{
  "components": {
    "schemas": {
...
      "BlockCreateRequest": {
        "properties": {
          "benchmark_every_n_weeks": {
            "anyOf": [
              {
                "maximum": 12.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": 4,
            "title": "Benchmark Every N Weeks"
          },
          "deload_every_n_weeks": {
            "default": 4,
            "maximum": 12.0,
            "minimum": 1.0,
            "title": "Deload Every N Weeks",
            "type": "integer"
          },
          "deload_volume_factor": {
            "default": 0.6,
            "exclusiveMinimum": 0.1,
            "maximum": 1.0,
            "title": "Deload Volume Factor",
            "type": "number"
          },
          "duration_weeks": {
            "default": 8,
            "maximum": 24.0,
            "minimum": 1.0,
            "title": "Duration Weeks",
            "type": "integer"
          },
          "goal": {
            "$ref": "#/components/schemas/BlockGoal"
          },
          "modality_mix": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Modality Mix",
            "type": "object"
          },
          "rationale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rationale"
          },
          "sessions_per_week": {
            "default": 3,
            "maximum": 7.0,
            "minimum": 1.0,
            "title": "Sessions Per Week",
            "type": "integer"
          },
          "start_date": {
            "format": "date",
            "title": "Start Date",
            "type": "string"
          },
          "weekly_template": {
            "items": {
              "$ref": "#/components/schemas/WeeklyTemplateSlot"
            },
            "title": "Weekly Template",
            "type": "array"
          }
        },
        "required": [
          "goal",
          "start_date"
        ],
        "title": "BlockCreateRequest",
        "type": "object"
      },
      "BlockGoal": {
        "enum": [
          "Strength",
          "Hypertrophy",
          "Power",
          "Hyrox",
          "CrossFit",
          "Running",
          "Calisthenics",
          "General",
          "Recomp"
        ],
        "title": "BlockGoal",
        "type": "string"
      },
      "BlockRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "deload_every_n_weeks": {
            "title": "Deload Every N Weeks",
            "type": "integer"
          },
          "deload_volume_factor": {
            "title": "Deload Volume Factor",
            "type": "number"
          },
          "duration_weeks": {
            "title": "Duration Weeks",
            "type": "integer"
          },
          "end_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "goal": {
            "$ref": "#/components/schemas/BlockGoal"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "modality_mix": {
            "additionalProperties": true,
            "title": "Modality Mix",
            "type": "object"
          },
          "rationale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rationale"
          },
          "sessions_per_week": {
            "title": "Sessions Per Week",
            "type": "integer"
          },
          "start_date": {
            "format": "date",
            "title": "Start Date",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/BlockStatus"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          },
          "weekly_template": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Weekly Template",
            "type": "array"
          }
        },
        "required": [
          "id",
          "user_id",
          "goal",
          "status",
          "start_date",
          "end_date",
          "duration_weeks",
          "sessions_per_week",
          "weekly_template",
          "modality_mix",
          "rationale",
          "deload_every_n_weeks",
          "deload_volume_factor",
          "created_at"
        ],
        "title": "BlockRead",
        "type": "object"
      },
      "BlockStatus": {
        "enum": [
          "active",
          "completed",
          "abandoned"
        ],
        "title": "BlockStatus",
        "type": "string"
      },
      "BlockUpdateRequest": {
        "properties": {
          "deload_volume_factor": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.1,
                "maximum": 1.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deload Volume Factor"
          },
          "modality_mix": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modality Mix"
          },
          "rationale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rationale"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BlockStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "BlockUpdateRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PlannedSessionRead": {
        "properties": {
          "benchmark_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Benchmark Key"
          },
          "block_id": {
            "title": "Block Id",
            "type": "integer"
          },
          "category": {
            "title": "Category",
            "type": "string"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "day_of_week": {
            "title": "Day Of Week",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_benchmark": {
            "title": "Is Benchmark",
            "type": "boolean"
          },
          "is_deload": {
            "title": "Is Deload",
            "type": "boolean"
          },
          "modality": {
            "title": "Modality",
            "type": "string"
          },
          "prescribed_content": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prescribed Content"
          },
          "scheduled_date": {
            "format": "date",
            "title": "Scheduled Date",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/SessionStatus"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          },
          "week_number": {
            "title": "Week Number",
            "type": "integer"
          },
          "workout_log_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workout Log Id"
          }
        },
        "required": [
          "id",
          "block_id",
          "user_id",
          "scheduled_date",
          "week_number",
          "day_of_week",
          "category",
          "modality",
          "status",
          "is_deload",
          "is_benchmark"
        ],
        "title": "PlannedSessionRead",
        "type": "object"
      },
      "PlannedSessionUpdateRequest": {
        "properties": {
          "scheduled_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled Date"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SessionStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "PlannedSessionUpdateRequest",
        "type": "object"
      },
...
      "SessionStatus": {
        "enum": [
          "pending",
          "completed",
          "skipped",
          "rescheduled"
        ],
        "title": "SessionStatus",
        "type": "string"
      },
...
      "TodaySessionResponse": {
        "properties": {
          "prescription": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prescription"
          },
          "session": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlannedSessionRead"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "session"
        ],
        "title": "TodaySessionResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WeeklyTemplateSlot": {
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "day_of_week": {
            "maximum": 7.0,
            "minimum": 1.0,
            "title": "Day Of Week",
            "type": "integer"
          },
          "modality": {
            "title": "Modality",
            "type": "string"
          }
        },
        "required": [
          "day_of_week",
          "category",
          "modality"
        ],
        "title": "WeeklyTemplateSlot",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/token"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "Performance Lab API",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/planning/blocks": {
      "get": {
        "operationId": "list_blocks_v1_planning_blocks_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BlockRead"
                  },
                  "title": "Response List Blocks V1 Planning Blocks Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Blocks",
        "tags": [
          "Planning"
        ]
      },
      "post": {
        "operationId": "create_block_v1_planning_blocks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlockCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlockRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Block",
        "tags": [
          "Planning"
        ]
      }
    },
    "/v1/planning/blocks/{block_id}": {
      "patch": {
        "operationId": "update_block_v1_planning_blocks__block_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "block_id",
            "required": true,
            "schema": {
              "title": "Block Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlockUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlockRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Block",
        "tags": [
          "Planning"
        ]
      }
    },
    "/v1/planning/sessions": {
      "get": {
        "operationId": "list_sessions_v1_planning_sessions_get",
        "parameters": [
          {
            "in": "query",
            "name": "start_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "in": "query",
            "name": "end_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Date"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlannedSessionRead"
                  },
                  "title": "Response List Sessions V1 Planning Sessions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Sessions",
        "tags": [
          "Planning"
        ]
      }
    },
    "/v1/planning/sessions/{session_id}": {
      "patch": {
        "operationId": "update_session_v1_planning_sessions__session_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlannedSessionUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlannedSessionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Session",
        "tags": [
          "Planning"
        ]
      }
    },
    "/v1/planning/today": {
      "get": {
        "operationId": "get_today_v1_planning_today_get",
        "parameters": [
          {
            "in": "query",
            "name": "goal",
            "required": false,
            "schema": {
              "default": "Strength",
              "title": "Goal",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodaySessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Today",
        "tags": [
          "Planning"
        ]
      }
    },
...
  }
}